Application specific file handling
It is possible to configure input sandbox files associated to an
application. Now it is no more necessary to specify these file in the
json input_sandbox field during job request (1st call). In case the
application file is present in the request, there are two possibilities.
If the flag override is true the user input will be ignored otherwise
the default value will be ignored. Anyhow the override can be still
ignored sending the file while finalizing the command POST/input